Reorganize code structure and add unit tests
Code reorganization:
- Move all .c files to src/ directory
- Move all .h files to include/ directory
- Update CMakeLists.txt with new paths
- Fix include paths for submodules (tiny-AES-c, tinycrypt)
- Fix compiler warnings (sign comparison, array bounds)

Unit testing infrastructure:
- Add Unity testing framework as git submodule
- Create test/ directory with CMakeLists.txt
- Add mock Pico SDK headers for host testing
- Create test.sh script for easy test execution
- Build tests on host (no Pico hardware needed)

Test coverage (test_macros.c):
- 13 comprehensive tests for macro parser/serializer
- Parsing: text, mnemonics, hex, ctrl shortcuts, explicit reports
- Mixed commands, comments, escaped quotes, multiple keydefs
- Serialization: text output, ctrl shortcuts
- Round-trip verification (parse → serialize → parse)
- Detailed keystroke verification in test_parse_simple_text

All tests passing:
- test_parse_simple_text (with detailed press/release verification)
- test_parse_mnemonic_trigger
- test_parse_hex_trigger
- test_parse_ctrl_shorthand
- test_parse_explicit_reports
- test_parse_mnemonic_commands
- test_parse_mixed_commands
- test_parse_with_comments
- test_parse_escaped_quotes
- test_parse_multiple_keydefs
- test_serialize_simple_text
- test_serialize_ctrl_shorthand
- test_roundtrip

+++ src/macros.c
@@ -391,7 +391,7 @@ static const char* keycode_to_mnemonic(uint8_t keycode) {
 
 // Reverse lookup: find ASCII char for a keycode (without modifier)
 static char keycode_to_ascii(uint8_t keycode, uint8_t modifier) {
-    for (int c = 0; c < sizeof(ascii_to_hid) / sizeof(ascii_to_hid[0]); c++) {
+    for (size_t c = 0; c < sizeof(ascii_to_hid) / sizeof(ascii_to_hid[0]); c++) {
         if (ascii_to_hid[c].key == keycode && ascii_to_hid[c].mod == modifier) {
             return (char)c;
         }
